Added Cart Api

+++ E-Commerce/E-Commerce.Domain/listResultDto/listResultDto.cs
@@ -9,7 +9,7 @@ namespace E_Commerce.Domain.listResultDto
     public class listResultDto<TEntity>
     {
         public IEnumerable<TEntity> entities { get; set; }
-        public int count { get; set; }
+        public decimal count { get; set; }
         public listResultDto()
         {
             entities = new List<TEntity>();
